AddRecipientsTest: switch to preSend()
The "dequeuing" of the IDN email addresses happens within `preSend()`, so these tests don't actually need to call `send()`.

The `PreSendTestCase` class, however, sets a `to` address already, while for this test class, we want each test to have a blank slate with regards to the `to`/`cc`/`bcc` addresses, so this class will be using the generic `TestCase` instead and only adding a `from` address to the property presets.